Deduplicate review-test fixtures; drop plan labels and history from comments
Shared stubs, the ReviewIssue helper and a selector factory now live in
ReviewLoopFixture; the per-file copies are gone. Comments lose plan labels
and change-history phrasing.
